feat(85-05): create EndgameOverallPerformanceSection 3-card composite, delete legacy files
- Create EndgameOverallPerformanceSection.tsx: 3-card layout (Games ending in
  Middlegame | At Endgame Entry | Endgame results) + Endgame Score Gap under
  Card 2 on desktop via lg:col-start-2
- Card 2 absorbs "Where you start" tile: entry-eval row + achievable score row
  with CR-01 OFFSET-form MiniBulletChart bounds preserved
- Delete EndgameGamesWithWithoutSection.tsx and EndgameStartVsEndSection.tsx
- Section question line replaces legacy section h3 + InfoPopover at section root
